fix: rename apis dir to ttrainer.
