+++ pkg/apis/trainer/v2alpha1/groupversion_info.go
