+++ pkg/apis/trainer/v2alpha1/trainjob_types.go
