+++ pkg/apis/trainer/v2alpha1/zz_generated.deepcopy.go
